requirements.txt
asgiref==3.8.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==5.1.3
django-environ==0.8.1
django-filter==24.3
gunicorn==20.1.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==5.3.0
